Separate modules Add README.md and TODO.md
