+++ README.md
@@ -1,0 +1,30 @@
+# The Do Programming Language
+
+Do is a Stack Based (Concatenative) Functional programming language.
+
+# Examples
+
+```
+4 5 + print
+6 2 - print
+4 5 * print
+2 8 / print
+3 5 % print
+
+2 3 4 + * print
+
+1 2 swap print print
+5 dup * print
+```
+prints
+```
+9
+-4
+20
+4
+2
+14
+1
+2
+25
+```
