+++ TODO.md
@@ -1,0 +1,7 @@
+# TODO
+
+- [] Diagnostics
+- [] Better Type Checking
+- [] Lists
+- [] Functions
+- [] REPL
