+++ resources/stack.do
@@ -1,0 +1,2 @@
+1 2 swap print print
+5 dup * print
